// src/app/models/unitSummary/unitSummaryResponse.model.ts
export class IUnitSummaryResponse {
    id: string;
    size: string;
    unit_class_kind: string;
    unit_type_produce_name: string;
    technology_level: string;
    equipment_count: string;
    equipment_max: string;
    equipment_quality: string;
    equipment_quality_required: string;
    equipment_wear: string;
    employee_count: string;
    employee_max: string;
    employee_required_by_equipment: string;
    employee_salary: string;
    city_salary: string;
    salary_cost: string;
    employee_level: string;
    city_education: string;
    employee_level_required: string;
    office_productivity: string;
    equipment_productivity: string;
    employee_productivity: string;
    manager_productivity: string;
    knowledge_area_id: string;
    knowledge_area_name: string;
    knowledge_area_kind: string;
    competence_value: string;
    competence_value_wo_bonus: string;
    all_staff: string;
    manager_level_required: string;
    productivity: string;
    loading: string;
    upgrade_time_to_finish: string;
    upgrade_size: string;
    upgrade_unit_type_employee_max: string;
    supply_warning: string;
    training: string;
    unit_type_name: string;
    unit_type_id: string;
    name: string;
    labor_max: string;
    labor_qty: string;
    speedup_cost: number;
    company_id: string;
    user_id: string;
    unit_type_produce_id: string;
    symbol: string;
    city_id: string;
    market_price: string;
    exclusive_market_price: string;
    auction_lot_id: string;
    is_bankrupt: string;
    on_bill: string;
    corp_id: string;
    office_id: string;
    on_holiday: string;
    is_student: string;
    unit_id: string;
    text: string;
    office_name: string;
    notice: string;
    already_learn: string;
}